Refactor: Modularize codebase for better maintainability
